change integer type to u32 in IV/EV screen

+++ raw/src/pokemon_summary_screen.c
@@ -135,7 +135,7 @@ static void SetMainMoveSelectorColor(u8 which)
 
 static void BufferIvOrEvStats(u8 mode)
 {
-    u8 hp, hp2, atk, def, spA, spD, spe;
+    u32 hp, hp2, atk, def, spA, spD, spe;
     u8 *currHPString = Alloc(20);
     const s8 *natureMod = gNatureStatTable[sMonSummaryScreen->summary.nature];
